fix: complete shell capture process extraction
